Reorganized safeguards code into more constant and helper files for better readability.
